changed GPU dense ColumnSlice to a constructor to avoid multiple creation of the storage object; minor cleanup w.r.t. ZeroInit() and ZeroValues()

+++ Examples/Extensibility/BinaryConvolution/BinaryConvolutionLib/BinaryConvolveOp.h
@@ -73,7 +73,7 @@ private:
         {
             auto numOutCols = !pad ? (w - size)/stride + 1 : (w - 1)/stride + 1;
             auto numOutRows = !pad ? (h - size)/stride + 1 : (h - 1)/stride + 1;
-            outputValue = MakeSharedObject<Value>(MakeSharedObject<NDArrayView>(DataType::Float, NDShape({ numOutRows , numOutCols, num_filters }), computeDevice));
+            outputValue = MakeSharedObject<Value>(MakeSharedObject<NDArrayView>(DataType::Float, NDShape({ (size_t)numOutRows, (size_t)numOutCols, (size_t)num_filters }), computeDevice));
         }
         
         // extract the output data
@@ -135,7 +135,7 @@ private:
         auto numOutCols = !pad ? (w - size)/stride + 1 : (w - 1)/stride + 1;
         auto numOutRows = !pad ? (h - size)/stride + 1 : (h - 1)/stride + 1;
         // return the appropriate output shape 
-        outputs.push_back(OutputVariable(NDShape({ numOutRows, numOutCols, num_filters }), leftOperand.GetDataType(), rightOperand.DynamicAxes()));
+        outputs.push_back(OutputVariable(NDShape({ (size_t)numOutRows, (size_t)numOutCols, (size_t)num_filters }), leftOperand.GetDataType(), rightOperand.DynamicAxes()));
         return OutputsVectorType(outputs);
     }
